Added toast messages for copy,move and rename folder functions.

+++ src/main/java/nl/bioinf/jp_kcd_wr/image_library/ui_commands/FileUICommandService.java
@@ -64,16 +64,22 @@ public class FileUICommandService implements UICommandService {
      * @author Kim Chau Duong
      */
     @Override
-    public void moveFile(String source, String destination) {
+    public boolean moveFile(String source, String destination) {
         Path pathFrom = getFullPath(source);
         Path pathTo = getFullPath(destination).resolve(new File(source).getName());
-        try{
-            logger.log(Level.INFO,"Moving directory from {0} to {1}", new Object[]{source, destination});
-            Files.move(pathFrom, pathTo, StandardCopyOption.REPLACE_EXISTING);
-            storageService.processExistingImageLibrary(new File(String.valueOf(getFullPath(destination))));
-
-        } catch (IOException e) {
-            logger.log(Level.WARNING, "{0} could not be moved", source);
+        if (Files.exists(pathTo)){
+            logger.log(Level.WARNING, "{0} already exists!", source);
+            return false;
+        } else {
+            try{
+                logger.log(Level.INFO,"Moving {0} from {1} to {2}", new Object[]{new File(source).getName(),source, destination});
+                Files.move(pathFrom, pathTo, StandardCopyOption.REPLACE_EXISTING);
+                storageService.processExistingImageLibrary(new File(String.valueOf(getFullPath(destination))));
+                return true;
+            } catch (IOException e) {
+                logger.log(Level.WARNING, "{0} could not be moved", source);
+                return false;
+            }
         }
     }
 
@@ -85,15 +91,22 @@ public class FileUICommandService implements UICommandService {
      * @author Kim Chau duong
      */
     @Override
-    public void copyFile(String source, String destination) {
+    public boolean copyFile(String source, String destination) {
         Path pathFrom = getFullPath(source);
         Path pathTo = getFullPath(destination).resolve(new File(source).getName());
-        try{
-            logger.log(Level.INFO,"Copying directory from {0} to {1}", new Object[]{source, destination});
-            FileUtils.copyDirectory(pathFrom.toFile(), pathTo.toFile());
-            storageService.processExistingImageLibrary(new File(String.valueOf(getFullPath(destination))));
-        } catch (IOException e) {
-            logger.log(Level.WARNING, "{0} could not be copied", source);
+        if (Files.exists(pathTo)){
+            logger.log(Level.WARNING, "{0} already exists!", source);
+            return false;
+        } else {
+            try{
+                logger.log(Level.INFO,"Copying {0} from {1} to {2}", new Object[]{new File(source).getName(),source, destination});
+                FileUtils.copyDirectory(pathFrom.toFile(), pathTo.toFile());
+                storageService.processExistingImageLibrary(new File(String.valueOf(getFullPath(destination))));
+                return true;
+            } catch (IOException e) {
+                logger.log(Level.WARNING, "{0} could not be copied", source);
+                return false;
+            }
         }
     }
 
@@ -103,16 +116,23 @@ public class FileUICommandService implements UICommandService {
      * @param renamedFileName new name
      */
     @Override
-    public void renameFile(String source, String renamedFileName) {
+    public boolean renameFile(String source, String renamedFileName) {
         Path oldPath = getFullPath(source);
         Path renamedPath = oldPath.resolveSibling(renamedFileName);
-        try{
-            logger.log(Level.INFO, "Renaming {0} to {1} in {2}", new Object[]{new File(source).getName(), renamedFileName, new File(source).getParent()});
-            Files.move(oldPath, renamedPath, StandardCopyOption.REPLACE_EXISTING);
-            storageService.processExistingImageLibrary(new File(String.valueOf(renamedPath)));
-        } catch (IOException e) {
-            logger.log(Level.WARNING, "{0} could not be renamed", source);
+        if (Files.exists(renamedPath)){
+            logger.log(Level.WARNING, "{0} already exists!", source);
+            return false;
         }
-
+        else {
+            try{
+                logger.log(Level.INFO, "Renaming {0} to {1} in {2}", new Object[]{new File(source).getName(), renamedFileName, new File(source).getParent()});
+                Files.move(oldPath, renamedPath, StandardCopyOption.REPLACE_EXISTING);
+                storageService.processExistingImageLibrary(new File(String.valueOf(renamedPath)));
+                return true;
+            } catch (IOException e) {
+                logger.log(Level.WARNING, "{0} could not be renamed", source);
+                return false;
+            }
+        }
     }
 }
